requirements.txt
mira-sdk==0.1.11
streamlit
python-dotenv
